requirements.txt
beautifulsoup4==4.12.2
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==41.0.3
Flask==2.3.2
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==4.9.3
MarkupSafe==3.0.2
pycparser==2.22
pycryptodome==3.21.0
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.31.0
simple-websocket==1.1.0
soupsieve==2.6
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
